remove unused code
