--- Picroup-iOS/Feedbacks/CreateImageState/CreateImageState+LocalStorage.swift
@@ -1,16 +1,0 @@
-//
-//  CreateImageState+LocalStorage.swift
-//  Picroup-iOS
-//
-//  Created by luojie on 2018/4/22.
-//  Copyright © 2018年 luojie. All rights reserved.
-//
-
-import Foundation
-import RxSwift
-import RxCocoa
-import RxFeedback
-
-extension DriverFeedback where State == CreateImageState {
-}
-
