+++ Picroup-iOS/Routers/Home/HomeViewController/HomeViewController.swift
@@ -16,9 +16,6 @@ import Apollo
 
 class HomeViewController: UIViewController {
     
-//    typealias Dependency = (state: Driver<HomeState>, events: (HomeState.Event) -> Void)
-//    var dependency: Dependency!
-    
     typealias Feedback = (Driver<MyInterestedMediaStateObject>) -> Signal<MyInterestedMediaStateObject.Event>
     @IBOutlet var presenter: HomeViewPresenter!
     
@@ -70,38 +67,6 @@ class HomeViewController: UIViewController {
             .debug("MyInterestedMediaStateObject.Event", trimOutput: true)
             .emit(onNext: store.on)
             .disposed(by: disposeBag)
-        
-        //        guard let (state, events) = dependency else { return }
-        //        typealias Feedback = (Driver<HomeState>) -> Signal<HomeState.Event>
-        
-        //        let uiFeedback: Feedback = bind(self) { (me, state) in
-        //            let _events = PublishRelay<HomeState.Event>()
-        //            let subscriptions = [
-        //                state.map { [Section(model: "", items: $0.items)] }.drive(me.presenter.items(_events))
-        //            ]
-        //            let events: [Signal<HomeState.Event>] = [
-        //                state.flatMapLatest {
-        //                    $0.shouldQueryMore ? me.presenter.collectionView.rx.isNearBottom.asSignal() : .empty()
-        //                    }.map { .onTriggerGetMore },
-        //                _events.asSignal(),
-        //            ]
-        //            return Bindings(subscriptions: subscriptions, events: events)
-        //        }
-        //
-        //        let vcFeedback: Feedback = bind(self) { (me, state)  in
-        //            let subscriptions = [
-        //                me.presenter.collectionView.rx.shouldHideNavigationBar()
-        //                    .emit(to: me.rx.setNavigationBarHidden(animated: true))
-        //            ]
-        //            let events: [Signal<HomeState.Event>] = [
-        //                .never(),
-        //                ]
-        //            return Bindings(subscriptions: subscriptions, events: events)
-        //        }
-        //
-        //        Signal.merge(uiFeedback(state), vcFeedback(state))
-        //            .emit(onNext: events)
-        //            .disposed(by: disposeBag)
         
         presenter.collectionView.rx.setDelegate(presenter)
             .disposed(by: disposeBag)
