+++ Picroup-iOS/Routers/Image/ImageDetailViewController/ImageDetailCell.swift
@@ -31,26 +31,6 @@ extension ImageDetailCell {
 
 extension ImageDetailCell.ViewModel {
     
-    init(imageDetailState state: ImageDetailState) {
-        let (item, meduim, starMedium) = (state.item, state.meduim, state.starMedium)
-        
-        let endAt = starMedium?.endedAt ?? meduim?.endedAt ?? item.endedAt
-        let remainTime = endAt.sinceNow
-        
-        self.imageViewMinioId = item.minioId
-        self.imageViewMotionIdentifier = item.id
-        self.progress = CGFloat(remainTime / 8.0.weeks)
-        self.lifeBarMotionIdentifier = "lifeBar_\(item.id)"
-        self.starButtonMotionIdentifier = "starButton_\(item.id)"
-        self.remainTimeLabelText = "\(Int(remainTime / 1.0.weeks)) 周"
-        self.commentsCountLabelText = "\(item.commentsCount) 条"
-        self.stared = (starMedium != nil) ? true : meduim?.stared
-        self.animatedChangeProgress = item.endedAt != endAt
-        
-        self.username = item.user.username
-        self.avatarId = item.user.avatarId
-    }
-    
     init(medium: MediumObject) {
         let remainTime = medium.endedAt.value?.sinceNow ?? 0
