+++ Picroup-iOS/Routers/Login/LoginViewController/LoginViewController.swift
@@ -15,9 +15,6 @@ import Apollo
 
 class LoginViewController: UIViewController {
     
-//    typealias Dependency = ([DriverFeedback<LoginState>.Raw]) -> Disposable
-//    let dependency: Dependency
-    
     fileprivate var loginViewPresenter: LoginViewPresenter!
     fileprivate typealias Feedback = (Driver<LoginStateObject>) -> Signal<LoginStateObject.Event>
 
@@ -71,34 +68,6 @@ class LoginViewController: UIViewController {
             .debug("LoginStateObject.Event", trimOutput: true)
             .emit(onNext: store.on)
             .disposed(by: disposeBag)
-        
-        //        guard let dependency = dependency else { return }
-        //        let (client, observer) = dependency
-        
-        //        loginViewPresenter = LoginViewPresenter(view: view)
-        //
-        //        let uiFeedback: DriverFeedback<LoginState>.Raw = bind(loginViewPresenter) { [snackbarController = snackbarController!] (presenter, state) in
-        //            let subscriptions = [
-        //                state.map { $0.next.username }.distinctUntilChanged().drive(presenter.usernameField.rx.text),
-        //                state.map { $0.next.password }.distinctUntilChanged().drive(presenter.passwordField.rx.text),
-        //                state.map { $0.shouldHideUseenameWarning }.distinctUntilChanged().drive(presenter.usernameField.detailLabel.rx.isHidden),
-        //                state.map { $0.shouldHidePasswordWarning }.distinctUntilChanged().drive(presenter.passwordField.detailLabel.rx.isHidden),
-        //                state.map { $0.isLoginButtonEnabled }.distinctUntilChanged().drive(presenter.raisedButton.rx.isEnabledWithBackgroundColor(.secondary)),
-        //                state.map { $0.trigger }.distinctUntilChanged().mapToVoid().drive(presenter.usernameField.rx.resignFirstResponder()),
-        //                state.map { $0.trigger }.distinctUntilChanged().mapToVoid().drive(presenter.passwordField.rx.resignFirstResponder()),
-        //                state.map { $0.user }.distinctUnwrap().map { _ in "登录成功" }.drive(snackbarController.rx.snackbarText),
-        //                state.map { $0.error }.distinctUnwrap().map { $0.localizedDescription }.drive(snackbarController.rx.snackbarText),
-        //            ]
-        //            let events = [
-        //                presenter.usernameField.rx.text.orEmpty.map(LoginState.Event.onChangeUsername),
-        //                presenter.passwordField.rx.text.orEmpty.map(LoginState.Event.onChangePassword),
-        //                presenter.raisedButton.rx.tap.map { LoginState.Event.onTrigger }
-        //            ]
-        //            return Bindings(subscriptions: subscriptions, events: events)
-        //        }
-        //
-        //        dependency([uiFeedback])
-        //            .disposed(by: disposeBag)
         
     }
 }
