+++ Picroup-iOS/Routers/Me/UserViewController/UserViewController.swift
@@ -86,29 +86,6 @@ class UserViewController: HideNavigationBarViewController {
             .emit(onNext: store.on)
             .disposed(by: disposeBag)
 
-//        let injectDependncy = self.injectDependncy(appStore: appStore)
-//        let uiFeedback = self.uiFeedback
-//        let queryUser = Feedback.queryUser(client: ApolloClient.shared)
-//        let queryMyMedia = Feedback.queryMyMedia(client: ApolloClient.shared)
-////        let showImageDetail = Feedback.showImageDetail(from: self)
-//        let pop = Feedback.pop(from: self)
-//
-//        let reduce = logger(identifier: "UserState")(UserState.reduce)
-//
-//        Driver<Any>.system(
-//            initialState: UserState.empty(userId: userId),
-//            reduce: reduce,
-//            feedback:
-//                injectDependncy,
-//                uiFeedback,
-//                queryUser,
-//                queryMyMedia,
-////                showImageDetail,
-//                pop
-//            )
-//            .drive()
-//            .disposed(by: disposeBag)
-        
         presenter.myMediaCollectionView.rx.shouldHideNavigationBar()
             .emit(onNext: { [weak presenter, weak self] in
                 presenter?.hideDetailLayoutConstraint.isActive = $0
@@ -117,37 +94,4 @@ class UserViewController: HideNavigationBarViewController {
             .disposed(by: disposeBag)
 
     }
-}
-extension UserViewController {
-    
-//    fileprivate func injectDependncy(appStore: AppStore) -> Feedback.Raw {
-//        return { _ in
-//            appStore.state.map { $0.currentUser?.toUser() }.asSignal(onErrorJustReturn: nil).map { .onUpdateCurrentUser($0) }
-//        }
-//    }
-    
-//    fileprivate var uiFeedback: Feedback.Raw {
-//        typealias Section = UserPresenter.Section
-//        return bind(presenter) { (presenter, state) -> Bindings<UserState.Event> in
-//            let meViewModel = state.map { UserViewModel(user: $0.user) }
-//            let subscriptions: [Disposable] = [
-//                meViewModel.map { $0.avatarId }.drive(presenter.userAvatarImageView.rx.imageMinioId),
-//                meViewModel.map { $0.username }.drive(presenter.displaynameLabel.rx.text),
-//                meViewModel.map { $0.username }.drive(presenter.usernameLabel.rx.text),
-//                meViewModel.map { $0.reputation }.drive(presenter.reputationCountLabel.rx.text),
-//                meViewModel.map { $0.followersCount }.drive(presenter.followersCountLabel.rx.text),
-//                meViewModel.map { $0.followingsCount }.drive(presenter.followingsCountLabel.rx.text),
-//                state.map { [Section(model: "", items: $0.myMediaItems)] }.drive(presenter.myMediaItems),
-//                ]
-//            let events: [Signal<UserState.Event>] = [
-//                state.flatMapLatest {
-//                    $0.shouldQueryMoreMyMedia ? presenter.myMediaCollectionView.rx.isNearBottom.asSignal() : .empty()
-//                    }.map { .onTriggerGetMoreMyMedia },
-//                presenter.myMediaCollectionView.rx.itemSelected.asSignal().map { .onTriggerShowImageDetail($0.item) },
-//                presenter.meBackgroundView.rx.tapGesture().when(.recognized).asSignalOnErrorRecoverEmpty().map { _ in .onPop }
-//                ]
-//            return Bindings(subscriptions: subscriptions, events: events)
-//
-//        }
-//    }
 }
