--- Picroup-iOS/States/LoginState.swift
@@ -1,101 +1,0 @@
-//
-//  LoginState.swift
-//  Picroup-iOS
-//
-//  Created by luojie on 2018/4/8.
-//  Copyright © 2018年 luojie. All rights reserved.
-//
-
-import Foundation
-
-private let minimalUsernameLength = 0
-private let minimalPasswordLength = 0
-
-struct LoginState: Mutabled {
-    
-    var next: LoginQuery
-    var user: UserDetailFragment?
-    var error: Error?
-    var trigger: Bool
-}
-
-extension LoginState {
-    var query: LoginQuery? {
-        return trigger ? next : nil
-    }
-    var logged: Bool { return user != nil }
-    var isUsernameValid: Bool { return next.username.count > minimalUsernameLength }
-    var isPasswordValid: Bool { return next.password.count > minimalPasswordLength }
-    var shouldLogin: Bool { return isUsernameValid && isPasswordValid && !trigger }
-}
-
-extension LoginState {
-    static func empty() -> LoginState {
-        return LoginState(
-            next: LoginQuery(username: "", password: ""),
-            user: nil,
-            error: nil,
-            trigger: false
-        )
-    }
-}
-
-extension LoginState: IsFeedbackState {
-    
-    enum Event {
-        case onSuccess(UserDetailFragment)
-        case onError(Error)
-        case onTrigger
-        case onChangeUsername(String)
-        case onChangePassword(String)
-    }
-}
-
-extension LoginState {
-    
-    static func reduce(state: LoginState, event: LoginState.Event) -> LoginState {
-        switch event {
-        case .onSuccess(let user):
-            return state.mutated {
-                $0.user = user
-                $0.error = nil
-                $0.trigger = false
-            }
-        case .onError(let error):
-            return state.mutated {
-                $0.user = nil
-                $0.error = error
-                $0.trigger = false
-            }
-        case .onTrigger:
-            guard state.shouldLogin else { return state }
-            return state.mutated {
-                $0.user = nil
-                $0.error = nil
-                $0.trigger = true
-            }
-        case .onChangeUsername(let username):
-            return state.mutated {
-                $0.next.username = username
-            }
-        case .onChangePassword(let password):
-            return state.mutated {
-                $0.next.password = password
-            }
-        }
-    }
-}
-
-enum LoginError: LocalizedError {
-    case usernameOrPasswordIncorrect
-}
-
-extension LoginError {
-    
-    var localizedDescription: String {
-        switch self {
-        case .usernameOrPasswordIncorrect: return "用户名或密码错误"
-        }
-    }
-}
-
